more error checks

+++ HazardObj.cpp
@@ -20,6 +20,8 @@
 #include "TRitem.h"
 #include "TS.h"
 #include "TrackItemObj.h"
+#include "ErrorMessage.h"
+#include "ErrorMessagesLib.h"
 
 HazardObj::HazardObj() {
     this->shape = -1;
@@ -85,6 +87,11 @@ void HazardObj::load(int x, int y) {
 void HazardObj::deleteTrItems(){
     TDB* tdb = Game::trackDB;
     TDB* rdb = Game::roadDB;
+
+    if(!checkForErrors()){
+        return;
+    }
+    
     for(int i = 0; i<this->trItemIdCount/2; i++){
         if(this->trItemId[i*2] == 0)
             tdb->deleteTrItem(this->trItemId[i*2+1]);
@@ -93,6 +100,45 @@ void HazardObj::deleteTrItems(){
         this->trItemId[i*2+1] = -1;
     }
 }
+
+bool HazardObj::checkForErrors(){
+    TDB* tdb = Game::trackDB;
+    TRitem* item = NULL;
+    for(int i = 0; i<this->trItemIdCount/2; i++){
+        if(this->trItemId[i*2] != 0)
+            continue;
+        item = tdb->trackItems[this->trItemId[i*2+1]];
+        
+        if(item == NULL) {
+            ErrorMessage *e = new ErrorMessage(
+                ErrorMessage::Type_Error, 
+                ErrorMessage::Source_World, 
+                QString("Object '") + type + "' - reference to trackItem not found. UiD: " + QString::number(UiD) + ". ",
+                        "World Object requires Interactive Item that does not exist in Track DB. \n"
+                );
+            e->setLocationXYZ(x, -y, position[0], position[1], -position[2]);
+            e->setObject((GameObj*)this);
+            ErrorMessagesLib::PushErrorMessage(e);
+            return false;
+        }
+        if(item->type != "hazzarditem"){
+            ErrorMessage *e = new ErrorMessage(
+                ErrorMessage::Type_Error, 
+                ErrorMessage::Source_World, 
+                QString("Object '") + type + "' - referenced trackItem has wrong type. UiD: " + QString::number(UiD) + ". ",
+                        QString("World Object doesn't match TDB data. Is World Database and TDB out of sync? \n")+
+                        "Expected: hazzarditem but found: "+item->type+" .\n"
+                        "This may cause fatal errors."
+                );
+            e->setLocationXYZ(x, -y, position[0], position[1], -position[2]);
+            e->setObject((GameObj*)this);
+            ErrorMessagesLib::PushErrorMessage(e);
+            return false;
+        }
+    }
+    return true;
+}
+
 
 void HazardObj::initTrItems(float* tpos){
     if(tpos == NULL)
